requirements.txt
selenium
webdriver-manager
pandas
google-cloud-bigquery
python-dotenv